Fixes after rebase.
